requirement.txt
django
djangorestframework
markdown
django-filter
beautifulsoup4
requests
psycopg2
lxml
selenium
fake-useragent